Moved the arenas-related classes to a sub-package.
