+++ SpectatorPlus/src/com/pgcraft/spectatorplus/SpectateAPI.java
@@ -1,5 +1,7 @@
 package com.pgcraft.spectatorplus;
 
+import com.pgcraft.spectatorplus.arenas.Arena;
+import com.pgcraft.spectatorplus.arenas.ArenasManager;
 import org.bukkit.Material;
 import org.bukkit.command.CommandSender;
 import org.bukkit.entity.Player;
