+++ SpectatorPlus/src/com/pgcraft/spectatorplus/SpectateCommand.java
@@ -6,6 +6,7 @@ import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.logging.Level;
 
+import com.pgcraft.spectatorplus.arenas.Arena;
 import org.apache.commons.lang.WordUtils;
 import org.bukkit.Bukkit;
 import org.bukkit.ChatColor;
