+++ SpectatorPlus/src/com/pgcraft/spectatorplus/SpectateCompleter.java
@@ -1,17 +1,18 @@
 package com.pgcraft.spectatorplus;
 
+import com.pgcraft.spectatorplus.arenas.Arena;
+import org.bukkit.Material;
+import org.bukkit.command.Command;
+import org.bukkit.command.CommandSender;
+import org.bukkit.command.TabCompleter;
+import org.bukkit.entity.Player;
+
 import java.text.Collator;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
-
-import org.bukkit.Material;
-import org.bukkit.command.Command;
-import org.bukkit.command.CommandSender;
-import org.bukkit.command.TabCompleter;
-import org.bukkit.entity.Player;
 
 public class SpectateCompleter implements TabCompleter {
